chore

+++ src/roq/binance_futures/market_data.hpp
@@ -42,7 +42,7 @@ class MarketData final : public web::socket::Client::Handler, public json::Marke
     virtual void operator()(Trace<StatisticsUpdate> const &, bool is_last) = 0;
   };
 
-  MarketData(Handler &, io::Context &, uint32_t stream_id, Shared &, size_t index);
+  MarketData(Handler &, io::Context &, uint16_t stream_id, Shared &, size_t index);
 
   MarketData(MarketData &&) = delete;
   MarketData(MarketData const &) = delete;
